add & optimization
1.优化业务逻辑，新增事件驱动
2.优化线程入线程池
3.新增redis缓存
4.新增权限管理
5.新增积分/等级分
6.新增打赏

+++ service/src/main/java/com/itellyou/service/article/impl/ArticleStarServiceImpl.java
@@ -1,20 +1,22 @@
 package com.itellyou.service.article.impl;
 
 import com.itellyou.dao.article.ArticleStarDao;
-import com.itellyou.model.sys.PageModel;
 import com.itellyou.model.article.ArticleInfoModel;
 import com.itellyou.model.article.ArticleStarDetailModel;
 import com.itellyou.model.article.ArticleStarModel;
-import com.itellyou.model.user.UserOperationalAction;
-import com.itellyou.model.user.UserOperationalModel;
-import com.itellyou.model.sys.EntityType;
-import com.itellyou.service.article.ArticleIndexService;
+import com.itellyou.model.sys.EntityAction;
+import com.itellyou.model.event.ArticleEvent;
+import com.itellyou.model.sys.PageModel;
 import com.itellyou.service.article.ArticleInfoService;
 import com.itellyou.service.article.ArticleSearchService;
-import com.itellyou.service.article.ArticleStarService;
+import com.itellyou.service.common.StarService;
+import com.itellyou.service.event.OperationalPublisher;
 import com.itellyou.service.user.UserInfoService;
-import com.itellyou.service.user.UserOperationalService;
 import com.itellyou.util.DateUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.cache.annotation.CacheConfig;
+import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.interceptor.TransactionAspectSupport;
@@ -22,27 +24,29 @@ import org.springframework.transaction.interceptor.TransactionAspectSupport;
 import java.util.List;
 import java.util.Map;
 
+@CacheConfig(cacheNames = "article")
 @Service
-public class ArticleStarServiceImpl implements ArticleStarService {
+public class ArticleStarServiceImpl implements StarService<ArticleStarModel> {
+
+    private Logger logger = LoggerFactory.getLogger(this.getClass());
 
     private final ArticleStarDao starDao;
     private final ArticleInfoService infoService;
     private final ArticleSearchService searchService;
-    private final ArticleIndexService indexService;
-    private final UserOperationalService operationalService;
     private final UserInfoService userService;
+    private final OperationalPublisher operationalPublisher;
 
-    public ArticleStarServiceImpl(ArticleStarDao starDao, ArticleInfoService infoService,ArticleSearchService searchService,ArticleIndexService indexService, UserOperationalService operationalService,UserInfoService userService){
+    public ArticleStarServiceImpl(ArticleStarDao starDao, ArticleInfoService infoService, ArticleSearchService searchService, UserInfoService userService, OperationalPublisher operationalPublisher){
         this.starDao = starDao;
         this.infoService = infoService;
         this.searchService = searchService;
-        this.indexService = indexService;
-        this.operationalService = operationalService;
         this.userService = userService;
+        this.operationalPublisher = operationalPublisher;
     }
 
     @Override
     @Transactional
+    @CacheEvict(key = "#model.articleId")
     public int insert(ArticleStarModel model) throws Exception {
         ArticleInfoModel infoModel = searchService.findById(model.getArticleId());
         try{
@@ -53,33 +57,35 @@ public class ArticleStarServiceImpl implements ArticleStarService {
             if(result != 1) throw new Exception("更新收藏数失败");
             result = userService.updateCollectionCount(model.getCreatedUserId(),1);
             if(result != 1) throw new Exception("更新用户收藏数失败");
+            operationalPublisher.publish(new ArticleEvent(this, EntityAction.FOLLOW,model.getArticleId(),infoModel.getCreatedUserId(),model.getCreatedUserId(), DateUtils.getTimestamp(),model.getCreatedIp()));
         }catch (Exception e){
+            logger.error(e.getLocalizedMessage());
             TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
             throw e;
         }
-        indexService.updateIndex(model.getArticleId());
-
-        operationalService.insertAsync(new UserOperationalModel(UserOperationalAction.FOLLOW, EntityType.ARTICLE,model.getArticleId(),infoModel.getCreatedUserId(),model.getCreatedUserId(), DateUtils.getTimestamp(),model.getCreatedIp()));
-
         return 1;
     }
 
     @Override
     @Transactional
-    public int delete(Long articleId, Long userId) throws Exception {
+    @CacheEvict(key = "#articleId")
+    public int delete(Long articleId, Long userId,Long ip) throws Exception {
+        ArticleInfoModel infoModel = searchService.findById(articleId);
         try{
+            if(infoModel == null) throw new Exception("错误的文章ID");
             int result = starDao.delete(articleId,userId);
             if(result != 1) throw new Exception("删除收藏记录失败");
             result = infoService.updateStars(articleId,-1);
             if(result != 1) throw new Exception("更新收藏数失败");
             result = userService.updateCollectionCount(userId,-1);
             if(result != 1) throw new Exception("更新用户收藏数失败");
+            operationalPublisher.publish(new ArticleEvent(this, EntityAction.UNFOLLOW,articleId,infoModel.getCreatedUserId(),userId, DateUtils.getTimestamp(),ip));
+
         }catch (Exception e){
+            logger.error(e.getLocalizedMessage());
             TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
             throw e;
         }
-        indexService.updateIndex(articleId);
-        operationalService.deleteByTargetIdAsync(UserOperationalAction.FOLLOW, EntityType.ARTICLE,userId,articleId);
         return 1;
     }
